runner-mcp/pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "vep-runner-mcp"
version = "0.1.0"
description = "CLI wrapper for running Ensembl VEP + plugins in a reproducible way"
readme = "README.md"
authors = [{ name = "Saleh Shekari" }]
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
  "typer[all]>=0.9.0",
  "click>=8.1.7,<9",
  "rich>=13.7,<14",
  "pydantic>=2.8,<3",
  "pyyaml>=6.0,<7",
  "pandas>=2.2.2,<2.3",
  "xxhash>=3.4,<4"
]

[project.scripts]
vep-run = "vep_runner_mcp.cli.run:cli"

[project.optional-dependencies]
dev = [
  "pytest>=8.2,<9",
  "pytest-cov>=5,<6",
  "ruff>=0.6,<0.7",
  "black>=24.8,<25",
]

[tool.hatch.build.targets.wheel]
packages = ["src/vep_runner_mcp"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["runner-mcp/src", "runner-mcp/tests"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q --maxfail=1"
testpaths = ["runner-mcp/tests"]
